% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciiToChar.R
\name{as.char}
\alias{as.char}
\title{as.char: An Alias For as.character()}
\usage{
as.char(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
as.char() is an alias for \code{\link{as.character}}().
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
