% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbindSafe.R
\name{cbindSafe}
\alias{cbindSafe}
\title{cbindSafe: Column-wise Concatenation of Objects With Potentially Different 
  Rows}
\usage{
cbindSafe(...)
}
\arguments{
\item{...}{two or more objects such as matrices, data.frames, arrays or vectors 
to concatenate along the second dimension (adding columns)}
}
\value{
Returns an object of the same class as the inputs (coerced to the same
   class if they differ) with additional columns, and the union of the rows.
}
\description{
cbindSafe() concatenates columns of arguments after ensuring same rows are 
present.
}
\details{
cbindSafe finds the union of (row) names of the objects.  It puts
the rows in common in the same order and applies \code{\link{cbind}},
then adds the rows unique to one object or the other, with missing values
for the object where the row is missing.
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
