% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle.R
\name{circle}
\alias{circle}
\title{Plot a Circle or Regular Polygon}
\usage{
circle(x, y = 0, add = T, segs = 100, how = "radius", ...)
}
\arguments{
\item{x}{A numeric vector of length 1 or 2:
\itemize{
   \item If length(x)==1: origin is assumed to be c(0,0); x-coordinate of 
      a radial segment.
   \item If length(x)==2: x-coordinates of origin (x[1]) and radius 
     endpoint.
}}

\item{y}{Optional numeric vector of length 1 or 2:
\itemize{
   \item If omitted: assume y-coordinates of zero.
   \item If length(y)==1: origin is assumed to be c(0,0); y-coordinate of 
     a radial segment.
   \item If length(y)==2: y-coordinates of origin (x[1]) and radius 
     endpoint.
}}

\item{add}{Logical value indicating whether circle should be added to an
existing plot (default) or to a new plot.}

\item{segs}{Integer; number of segments used to approximate a circle.  A small number
will produce a regular polygon inscribed within the radius, with a vertex
at x, y}

\item{how}{Character string specifying how x and y relate to the circle geometry.
Currently only radius is implemented.}

\item{...}{Other named arguments passed to \code{\link[graphics]{lines}} or 
\code{\link[graphics]{plot}} (such as lwd, lty, col, etc.)}
}
\value{
Returns a data.frame with columns x and y specifying coordinates of 
the vertices.
}
\description{
Plot a circle or regular polygon of a given radius at a given location.
}
\seealso{
\code{\link[graphics]{par}} and \code{\link[graphics]{plot.default}} give
  more detailed descriptions of the ... plotting parameters.
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
