% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enclose.plot.R
\name{enclose.plot}
\alias{enclose.plot}
\title{Draw Edge Around A Plot.}
\usage{
enclose.plot(...)
}
\arguments{
\item{...}{arguments to be passed to \code{\link[graphics]{abline}}()-- col, lwd, 
lty, etc.}
}
\value{
Returns a named numeric vector with left, right, bottom and top limits of 
  the current plot; same as returned by par("usr"), but with named elements
}
\description{
Enclose a plot with lines at the axes; useful for the barplot function.
}
