% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedTickLabels.R
\name{groupedTickLabels}
\alias{groupedTickLabels}
\title{Label Axes With Hierarchically Organized Groups}
\usage{
groupedTickLabels(
  groups,
  side,
  lwd = 1,
  col = "black",
  cex = 0.7,
  las = 0,
  at,
  gaps,
  plot.labels = TRUE,
  group.tix = TRUE,
  meta.label = TRUE
)
}
\arguments{
\item{groups}{A data.frame() of annotations that define groupings based on their
values, orderd from most general to most specific.  The ordering of 
the rows must correspond to the ordering of the items plotted along 
the axis which will be labeled.}

\item{side}{Which axis to label-- see \code{\link[graphics]{axis}}.}

\item{lwd}{Vector of line widths for grid lines dividing groups, with each
element corresponding a column in groups.  If NA or omitted, no line
will be drawn for a given grouping.}

\item{col}{Vector of colors for the lines between groupings.}

\item{cex}{Numeric vector specifying sizing of labels for each annotation in
groups.  The last element will be reused for subsequent annotations.
Labels will not be printed if cex is zero or NA for a given 
annotation.}

\item{las}{A vector of integers in the range 0 to 2 indicating how the labels
for each annotation in groups should be oriented.  If fewer elements
are supplied than annotations (ncol(groups)), the remaining labels
will be placed perpendicular to the axis.}

\item{at}{Numeric vector specifying centers along the axes, to be used when 
spacing not uniform or separated values other than one.  By default,
at will be set at 1:nrow(groups).}

\item{gaps}{Numeric vector, specifying gap to leave, specified as margin lines,
between successive group labels; default is rep(0.5,ncol(groups)).
The last value specifies the gap between the axis line and the group
closest to the axis (i.e., the last group).}

\item{plot.labels}{Logical; if FALSE, do not add group labels to the plot, but
return placement info and print suggested margin.}

\item{group.tix}{Logical; if TRUE, draw lines connecting tick marks within 
each group, for each grouping variable.}

\item{meta.label}{Logical; should variable labels be plotted at the left edge
or top?}
}
\value{
Returns a list with elements named according to the columns of groups,
      with elements:
   \itemize{
     \item \strong{labels} Character vector with labels for the unique 
        groupings at this level of the hierarchy (column of groups).
     \item \strong{N} Numeric vector, same length as labels, with the 
        number of items (rows of groups) within this group. 
     \item \strong{mids} Numeric vector, same length as labels, with the 
        midpoints of each group, used to center labels along the axis.
     \item \strong{breaks} Numeric vector, length(labels)+1, with midpoints
       between successive groups (including before first and after last).
     \item \strong{outerLine}, {innerLine}: numeric values specifying 
       position of labels and grouping lines away from the axis
     \item \strong{group} Numeric vector, length equal to n.rows(groups),
       with the group number of each row of groups.  
     \item \strong{par} Vector of graphics parameters applied to this level 
       of the hierarchy, with names "line", "cex", "las", "adj", "padj".
   }
}
\description{
groupedTickLabels() labels and divides groups along the x- or y-axis.
}
\details{
This function is useful when the data along one axis is ordered by 
hierarchically related groups. The most general groups are labeled farthest
from the axis, and by default have larger labels and heavier lines between
them than more specific subgroups.

The user must provide sufficient margin space along the axis for the labels
via the par("mar") value for the side to be labelled.  A suggested value may
be generated by calling the function with plot.labels=FALSE.

Spacing required by the labels is determined primarily by the las and cex 
arguments, as well as gaps.  In general, you may have to play with the 
arguments a bit to make the labels look readable.
}
\seealso{
\code{\link[graphics]{par}} for more details about plotting 
  parameters lwd, col, cex and las.
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
