% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{Image}
\alias{Image}
\title{An Alternative Image Plotting Function}
\usage{
Image(z, xaxis, yaxis, NA.color = "gray", ...)
}
\arguments{
\item{z}{A numeric matrix or data.frame containing image data.}

\item{xaxis, yaxis}{Lists of arguments to the \code{\link[graphics]{axis}}. All arguments 
should be explicitly named.  The "side" argument may be omitted. 
An example:

\code{   xaxis=list(side=1, at=1:4, labels=LETTERS[1:4], cex.axis=0.8)}}

\item{NA.color}{Color for missing values.  Set to NA for transparency (so the color 
specified by \code{\link[graphics]{par}}(bg="") will show through.  Note
that this is accomplished by plotting a rectangle over the whole plot of 
the specified color, then replotting the image, so if you are adding the 
image to an existing plot, you may want to set this argument to NA}

\item{...}{Other arguments passed to \code{\link[graphics]{image}}.}
}
\description{
Image is a modified version of the \code{\link[graphics]{image}} function 
that behaves more like Matlab imagesc.
}
\details{
Image behaves like \code{\link[graphics]{image}}, with the following 
differences:
\itemize{
  \item The z argument may be a data.frame with numeric columns; it will be
    automatically converted to a matrix.
  \item The data matrix z is transposed, so that columns run along the x-axis 
    and rows along the y-axis.
  \item The direction of the y-axis is reversed, so that zero is at the top.
  \item If omitted, the x and y arguments are assumed from the size of z, 
     i.e.,
     
         \code{   x <- 0.5 + (0:ncol(z))}
         
         \code{   y <- 0.5 + (0:nrow(z))}
         
      If supplied, the x and y arguments must be explicitly named.
  \item If a zlim argument is supplied, values outside the zlim range are 
    replaced with the appropriate limit (floor or ceiling values to zlim).
  \item You may annotate the axes using the xaxis and yaxis list arguments 
}
  
This behavior should be familiar to users of Matlab's imagesc function.
}
\seealso{
\code{\link[graphics]{image}}, \code{\link[graphics]{par}}.
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
