% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciiToChar.R
\name{is.char}
\alias{is.char}
\title{is.char: An Alias For is.character()}
\usage{
is.char(x)
}
\arguments{
\item{x}{object to be coerced or tested.}
}
\description{
is.char() is an alias for \code{\link{is.character}}().
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
