% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_README.R
\name{make_README}
\alias{make_README}
\title{Generates README.md from *.rmd package documentation.}
\usage{
make_README(pkg.R_file, out.file)
}
\arguments{
\item{pkg.R_file}{character value specifying file path/name to be converted.}

\item{out.file}{optional character value specifying file path/name for output *.md file.}
}
\value{
Invisibly returns a character vector with the output, one line per row.
}
\description{
make_README() is a utility to generate a README.md file from microRutils.R
by converting the R markdown to GitHub-friendly markdown.
}
\details{
Might be reusable, but meant for internal use (Not exported). Caveat emptor; 
hasn't been tested much.  If you really want to try it, use 
microRutils:::make_README.R().
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
