% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonUniq.R
\name{nonUniq}
\alias{nonUniq}
\title{nonUniq: Find Replicated Elements In A Vector}
\usage{
nonUniq(x, value = F, na.rm = T)
}
\arguments{
\item{x}{vector or matrix. Missing values are ignored.}

\item{value}{logical; if TRUE, return the unique elements in x that occur more
than once.  If FALSE (default), return their indices.}

\item{na.rm}{logical; if TRUE, ignore missing values (NA's)}
}
\value{
If value==F, an integer vector with positions (indices) of non-unique 
   elements of x; otherwise, a vector with the unique elements of x that
   occur more than once.  Thus:
      all(nonUniq(x, value=T)==unique(x[nonUniq(x)])) is TRUE
}
\description{
nonUniq(x) returns items in x that occur more than once (indices or values)
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
