% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberReplicates.R
\name{numberReplicates}
\alias{numberReplicates}
\title{numberReplicates: Number Recurring Elements}
\usage{
numberReplicates(x, number.NAs = FALSE)
}
\arguments{
\item{x}{a vector that includes non-unique elements}

\item{number.NAs}{logical; if TRUE, missing values will be numbered; otherwise,
return an NA for each missing value}
}
\value{
returns a numeric vector where replicates of each unique value in x are 
   numbered from 1 to the number of replicates found in the order in which 
   they are found in x.  The replicate numbers returned correspond to the 
   elements of x as it was passed in.
}
\description{
numberReplicates numbers instances of each non-unique value in the order 
found
numberReplicates
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
