% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cString.R
\name{orString}
\alias{orString}
\title{orString: Generate An OR Regular Expression Substring}
\usage{
orString(chars)
}
\arguments{
\item{chars}{a character vector of strings to be concatenated}
}
\value{
a character string, of the form "(a|b|c)",  where a, b and c are the
   elements of chars, concatentated
}
\description{
orString concatenates strings for use in regular expression, like "(a|b|c)"
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
