% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.print.R
\name{p.print}
\alias{p.print}
\title{Formats p-values for readability.}
\usage{
p.print(p, llim, ...)
}
\arguments{
\item{p}{A numeric vector of values between zero and one.}

\item{llim}{A numeric value between zero and one (exclusive) specifying the lowest
p-value to report.  For example, if llim=0.001, values below this will
be printed as "<0.001".}

\item{...}{Other argments passed to \code{\link[base]{format}}().}
}
\value{
Returns a character vector of p-values.  Values >= than 0.1 are reported 
  with two significant figures; values <0.1 are reported with one significant
  figure.
}
\description{
p.print() formats p-values for readability, with values floored at a lower
limit.
}
\seealso{
\code{\link[base]{format}}.
}
