% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent}
\alias{percent}
\title{percent: Format Numeric Values As (Character) Percentages}
\usage{
percent(fractions, digits, justify = F)
}
\arguments{
\item{fractions}{numeric vector, matrix or array, where one corresponds to 
100%.  Missing values are permitted.  Use lapply() to format numeric
columns of a data.frame.}

\item{digits}{number of digits to the right of the decimal place to preserve.
If missing, defaults to the minimum number of digits such that all
results greater than zero will be represented as greater than zero.}

\item{justify}{logical, should the results be right justified by padding with 
spaces?}
}
\value{
returns a character vector with same dimensions as fractions, where values
   are represented as percentages, including "%" symbol.  Infinite values
   are returned as "Inf" or "-Inf"; missing values as NA's (unquoted).
}
\description{
percent() reformats a (numeric) fraction as a (character) percentage
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
