% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2Dhist.R
\name{plot2Dhist}
\alias{plot2Dhist}
\title{Generate and Plot a 2D-Histogram as a Heatmap}
\usage{
plot2Dhist(
  xdata,
  ydata,
  breaks = NULL,
  xbreaks = NULL,
  ybreaks = NULL,
  base = 1,
  col,
  xlab = NULL,
  ylab = NULL,
  plot.zero = 0,
  ...
)
}
\arguments{
\item{xdata, ydata}{Numeric vectors or matrices of the same size, with data to be
plotted as a 2D-histogram.  Data may also be passed in with xdata only, 
as a two-column array, data.frame or matrix with numeric values, but in
this case, \strong{ALL SUBSEQUENT ARGUMENTS MUST BE NAMED.}}

\item{breaks, xbreaks, ybreaks}{Same as breaks argument to \code{\link[graphics]{hist}}() function.  If
breaks is supplied, xbreaks and ybreaks are ignored and the same breaks
are applied to each axis.  Otherwise, xbreaks and ybreaks may be 
specified separately.  If omitted, breaks are determined for each 
variable as per the hist() function.}

\item{base}{Number specifying base for the z-axis.  Default of 1 plots the 
intensities in each bin, untransformed.  If a value greater than one is
supplied, the z-axis is transformed as log(intensities+1, base).}

\item{col}{Vector of colors specifying colormap to use.  See, for example,
\code{\link{hot}}.}

\item{xlab, ylab}{Labels for the x- and y-axes.  If omitted, the names of the 
variables used for the xdata and ydata arguments will be used.}

\item{plot.zero}{Value to plot zero count bins. Use NA to make those cells with 
zero counts behave as NA's (i.e. having no color; see help for 
\code{\link[graphics]{image}}).}

\item{...}{Additional arguments passed to \code{\link[graphics]{image}}}
}
\value{
Returns a list object with named elements:
   \itemize{
     \item \strong{xbreaks, ybreaks} 
        The breaks specifying binning of x- and y-data.
     \item \strong{xmids, ymids} Midpoints of the bins.
     \item \strong{intensities} Matrix with number of counts in each bin.
     \item \strong{zdata} Matrix of z-axis data as plotted (i.e.,  
        log transformed if base!=1).
     \item \strong{xdata, ydata, base} Same as input arguments.
   }
}
\description{
plot2Dhist generates a two-dimensional histogram image.
}
\details{
This function calculates a histogram of the join distribution of two 
variables, and plots the resulting matrix as a heatmap with the first 
variable along the x axis an the second along the y, with values increasing
from the lower left corner.  The axes are presented in the units of the 
input data (i.e., not bin numbers, but their values.)
}
\seealso{
\code{\link[graphics]{hist}}, \code{\link[graphics]{image}}.
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
