% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order.R
\name{pvalOrder}
\alias{pvalOrder}
\title{pvalOrder: Order Values By Levels, Then Numeric Value}
\usage{
pvalOrder(pvals, p.cut = 0.05)
}
\arguments{
\item{pvals}{numeric vector of values between zero and one}

\item{p.cut}{numeric vector of significance thresholds}
}
\description{
Order a set of p-values first on significance level, then on their values.
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
