% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbindSafe.R
\name{rbindSafe}
\alias{rbindSafe}
\title{rbindSafe: Row-wise Concatenation of Objects With Potentially Different 
  Columns}
\usage{
rbindSafe(...)
}
\arguments{
\item{...}{Two or more objects such as matrices, data.frames, arrays or vectors 
to concatenate along the first dimension (adding rows).}
}
\value{
Returns an object of the same class as the inputs (coerced to the same
   class if they differ) with additional rows, and the union of the columns.
}
\description{
rbindSafe() concatenates rows of arguments after ensuring same columns are 
present
}
\details{
rbindSafe finds the union of (column) names of the objects.  It puts
the columns in common in the same order and applies \code{\link{rbind}},
then adds the columns unique to one object or the other, with missing values
for the object where the column is missing.
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
