% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomColormaps.R
\name{red.green}
\alias{red.green}
\alias{yellow.blue}
\alias{grayscale}
\alias{jet}
\alias{purple.orange}
\alias{aqua.brown}
\alias{purple.green}
\alias{hot}
\alias{hot2}
\alias{orange.blue}
\title{Custom Colormaps For Heat Maps}
\usage{
red.green(levels = 100, space = "rgb", thru = "black", ...)

yellow.blue(levels = 100, space = "rgb", ...)

grayscale(levels = 100, space = "rgb", ...)

jet(levels = 100, space = "rgb", ...)

purple.orange(levels = 100, space = "rgb", ...)

aqua.brown(levels = 100, space = "rgb", ...)

purple.green(levels = 100, space = "rgb", ...)

hot(levels = 100, space = "rgb", ...)

hot2(levels = 100, space = "rgb", ...)

orange.blue(levels = 100, space = "rgb", thru = "black", ...)
}
\arguments{
\item{levels}{Integer, number of levels in the color map}

\item{space}{a character string; interpolation in RGB or CIE Lab
    color spaces.}

\item{thru}{A color, either named or as \code{\link[grDevices]{rgb}}() output, that
will serve as midpoint for color maps that are symmetric about a central
value.}

\item{...}{arguments to pass to \code{colorRamp}.}
}
\value{
A character vector with elements of 7 or 9 characters, "#" followed by the 
  red, blue, green and optionally alpha values in hexadecimal (after 
  rescaling to 0 ... 255). The optional alpha values range from 0 (fully
  transparent) to 255 (opaque).
}
\description{
A set of functions that to define various attractive color scales
}
\details{
Many of these functions (\code{orange.blue}, \code{yellow.blue},
\code{red.green}, \code{purple.orange}, \code{aqua.brown} and
\code{purple.green}) interpolate between three colors, where the function
name contains the start and stop colors, and the "thru" argument is used to
define a midpoint (typically black or white).  Many of these were defined to
accomodate people with color blindness.

Colormap functions that do not follow this convention are:
\itemize{
  \item \strong{grayscale} maps shades of grey from black to white.
  \item \strong{hot} mimics black-body radiation, passing from black to white
     through red, orange and yellow (like iron being heated up).
  \item \strong{hot2} is similar to hot, but has a more subdued shard of red
    as the first value after black. This is useful for de-emphasizing single
    data points in a heatmap of a 2D-histogram.
  \item \strong{jet} interpolates from dark blue to dark red through shades
     of royal blue, cyan, green, yellow and orange.
}
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
