% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowZs.R
\name{rowMads}
\alias{rowMads}
\title{rowMads: Fast Calculation of MADs by Row.}
\usage{
rowMads(X, na.rm = T, constant = 1.4826)
}
\arguments{
\item{X}{Numeric matrix, or object that can be coerced to matrix.}

\item{na.rm}{Logical; should NA's be omitted?}

\item{constant}{Scale factor; numeric constant for asymptotically normal consistency.}
}
\description{
rowMads(X) returns median absolute deviation (MAD) of each row of matrix x.
}
\seealso{
\code{\link[stats]{mad}}.
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
