% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowZs.R
\name{rowSds}
\alias{rowSds}
\title{rowSds: Fast Calculation of Standard Deviations of Rows}
\usage{
rowSds(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric matrix-like object}

\item{na.rm}{Logical; should NA's be omitted prior to calculations?}
}
\value{
Returns a numeric vector with length equal to number of rows of x.
}
\description{
rowSds(x) calculates standard deviatons across rows, faster than 
\code{apply(x, 1, sd)}
}
