% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectRows.R
\name{selectRows}
\alias{selectRows}
\title{selectRows: Row Selection of A Data.Frame}
\usage{
selectRows(meta, criteria, rownames = TRUE, not = c(), reorder = FALSE)
}
\arguments{
\item{meta}{a data.frame from which to select rows}

\item{criteria}{a list of named vectors, where the names correspond to columns
of meta, and the elements are values of those columns to include.  NA
is a permissible value.  Logically, the values within each element of 
criteria act as an OR; values between elements are treated as an AND.}

\item{rownames}{logical; if TRUE, rownames are returned rather than numeric 
indices}

\item{not}{character vector specifying names of criteria where only rows of meta
that do NOT include the elements specified in the criteria list}

\item{reorder}{logical; if TRUE, reorder the row indices by the values of the
columns specified by criteria, applied in order}
}
\value{
Returns a vector of numeric row indices (rownames==FALSE) or character
    rownames of meta that matched the criteria specified
}
\description{
selectRows returns row indices of a data.frame matching column-value criteria
}
\details{
The criteria are applied in order, so this will run faster if the most
exclusive criteria are applied first.  However, the reorder option orders
the indices according to the order of the criteria annotations
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
