% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strrev.R
\name{strrev}
\alias{strrev}
\title{strrev: Reverse Character Order}
\usage{
strrev(strs)
}
\arguments{
\item{strs}{A vector, matrix or array of character values, or other data type
that can be coerced to type character.  Note that if strs is numeric,
trailing zeros will not be preserved.}
}
\value{
A vector, matrix or array of character values with the same dimensions
    and other attributes (i.e., names) as the input argument, but with the
    the order of characters in the individual elements reversed.
}
\description{
strrev() reverses the order of characters within each element of a character 
   vector
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
