% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.R
\name{sym}
\alias{sym}
\title{sym: Get A Numeric Range That Is Symmetric Around Zero}
\usage{
sym(lims = 1, qtile = 1)
}
\arguments{
\item{lims}{A numeric value or vector.  If its length is greater than one, the
quantile specified by the qtile argument will be calculated for 
the absolute values of lims.}

\item{qtile}{A numeric value in the range 0-1 (inclusive).  If lims contains
multiple values, this quantile of the distributions of abs(lims)
will be used to define the interval.  The default value of 1 will 
return max(abs(lims)).}
}
\value{
Returns a symmetric interval about zero, from -lims to lims, or if lims
   contains more than one value, a symmetric interval about zero that 
   contains the fraction of the data specified by qtile.
}
\description{
sym returns a symmetric numeric interval, centered about zero.
}
\details{
This function is useful for setting symmetric limits on graphs (among other
things.)
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
