% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untab.R
\name{untab}
\alias{untab}
\title{untab: Convert Table To Array}
\usage{
untab(a.table)
}
\arguments{
\item{a.table}{an object of class table, such as generated by the table() or
xtab() functions}
}
\value{
Returns an array with the same dimensions and dimnames as a.table, that is
   NOT a table object, and may thus be manipulated like an array.
}
\description{
function to convert a table object to an array object
}
\details{
Table objects tend to get automatically converted into data.frames by
functions.  These data.frames contain columns with factor levels and
frequencies that get in the way.  as.array() doesn't do anything to tables,
because tables are already arrays.  This function gets around these issues.
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
