% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.pt.R
\name{x.pt}
\alias{x.pt}
\alias{y.pt}
\title{Get the Value Associated With A Fraction of X- or Y-Axis}
\usage{
x.pt(fraction)

y.pt(fraction)
}
\arguments{
\item{fraction}{Numeric values between zero and one specifying distance from the left (x)
or bottom (y) limit of the current plot relative to its range.}
}
\value{
Returns numeric value of specified location in the units of the axis.
}
\description{
Returns the coordinate of a point that is a given fraction between the actual
axis limits of the current plot, in the units of that axis' data.
}
