% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yyyymmdd.R
\name{yyyymmdd}
\alias{yyyymmdd}
\title{yyyymmdd: Get A Time Stamp}
\usage{
yyyymmdd(hms = FALSE, sep = "")
}
\arguments{
\item{hms}{logical; if TRUE append hour, minutes and seconds}

\item{sep}{character value to separate date and time if hms is TRUE}
}
\value{
Returns the current data (and time if hms==TRUE) as a character value
       formatted as YYYYMMDD(_HHMMSS).
}
\description{
yyyymmdd() returns current date, and optionally time, in YYYYMMDD format
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
